Move playerTracking hook to worker
